- Added Agent-Credentials
  - Updated the Credentials component to support Agent-Credentials by allowing user to create global Credentials, or create Agent-Credentials from an existing global value.
  - New Credentials API :
       - /AddAgentCredential (Adds a new Credential with an AgentId)
       - / GetCredentialByName/{credentialName} (Gets the Credential with the specified name. Retrieves any existing agent-credentials if the request is made by the agent )

- Updated /RaiseBusinessEvents to support additional fields (Message and Payload)

Related work items: #917, #918

+++ OpenBots.Server.Model/Job/Job.cs
@@ -11,7 +11,7 @@ namespace OpenBots.Server.Model
         public Guid? AgentGroupId { get; set; }
         public DateTime? StartTime { get; set; }
         public DateTime? EndTime { get; set; }
-        public double? ExecutionTimeInMinutes { get; set; }
+        public long? ExecutionTimeInMinutes { get; set; }
         public DateTime? EnqueueTime { get; set; }
         public DateTime? DequeueTime { get; set; }
         [Required]
